Added the feature TimeExtender in Auto Update mode (#3064)
This feature allow the user(if activate in the deck config) to have a
opportunity to increase the time before the auto-update of the
question/answer. At half of the auto-update time a button appears and if
clicked more time is added before the auto update of the
question/answer.